fix cast for format
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gworkspace/trunk@39154 72102866-910b-0410-8b05-ffd578937521

+++ GWorkspace/GWorkspace.m
@@ -1217,7 +1217,7 @@ static GWorkspace *gworkspace = nil;
     NSString *msg2 = NSLocalizedString(@"items?", @"");
   
     if (NSRunAlertPanel(nil,
-                [NSString stringWithFormat: @"%@ %i %@", msg1, count, msg2],
+                        [NSString stringWithFormat: @"%@ %lu %@", msg1, (unsigned long)count, msg2],
                 NSLocalizedString(@"Cancel", @""),
                 NSLocalizedString(@"Yes", @""),
                 nil)) {
@@ -2653,7 +2653,7 @@ static GWorkspace *gworkspace = nil;
       NSString *msg2 = NSLocalizedString(@"items?", @"");
 
       if (NSRunAlertPanel(nil,
-                  [NSString stringWithFormat: @"%@ %i %@", msg1, count, msg2],
+                          [NSString stringWithFormat: @"%@ %lu %@", msg1, (unsigned long)count, msg2],
                   NSLocalizedString(@"Cancel", @""),
                   NSLocalizedString(@"Yes", @""),
                   nil)) {
